# docs/swagger.yaml
definitions:
...
  go-enterprise-api_internal_domain_gatewayapi.GatewayAPI:
    properties:
      base_path:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_checked_at:
        type: string
      methods:
        items:
          type: string
        type: array
      name:
        type: string
      note:
        type: string
      rate_limit_rpm:
        type: integer
      requires_auth:
        type: boolean
      status:
        type: string
      updated_at:
        type: string
      upstream:
        type: string
    type: object
...
  go-enterprise-api_internal_infrastructure_gatewaymonitor.Entry:
    properties:
      bytes:
        type: integer
      ip:
        type: string
      method:
        type: string
      rt:
        type: number
      status:
        type: integer
      time:
        type: string
      ua:
        type: string
      upstream:
        type: string
      uri:
        type: string
    type: object
  go-enterprise-api_internal_infrastructure_gatewaymonitor.TodayStats:
    properties:
      2xx:
        type: integer
      4xx:
        type: integer
      5xx:
        type: integer
      count:
        type: integer
      errors:
        type: integer
      latency_ms:
        type: integer
    type: object
  go-enterprise-api_internal_interface_httpapi.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  go-enterprise-api_internal_interface_httpapi.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorBody'
    type: object
...
  internal_interface_http_handler.createGatewayAPIRequest:
    properties:
      base_path:
        type: string
      is_active:
        type: boolean
      methods:
        items:
          type: string
        type: array
      name:
        type: string
      note:
        type: string
      rate_limit_rpm:
        type: integer
      requires_auth:
        type: boolean
      upstream:
        type: string
    type: object
...
  internal_interface_http_handler.gatewayStatsResponse:
    properties:
      recent:
        items:
          $ref: '#/definitions/go-enterprise-api_internal_infrastructure_gatewaymonitor.Entry'
        type: array
      today:
        $ref: '#/definitions/go-enterprise-api_internal_infrastructure_gatewaymonitor.TodayStats'
    type: object
...
  internal_interface_http_handler.updateGatewayAPIRequest:
    properties:
      base_path:
        type: string
      is_active:
        type: boolean
      methods:
        items:
          type: string
        type: array
      name:
        type: string
      note:
        type: string
      rate_limit_rpm:
        type: integer
      requires_auth:
        type: boolean
      upstream:
        type: string
    type: object
...
host: localhost:18080
info:
  contact:
    name: GitOps
  description: 'Enterprise boilerplate: Dynamic Rate Limit, Soft Delete, Audit Logs,
    RBAC, JWT + Redis sessions, Microsoft Entra (OIDC). All requests pass through
    the Nginx gateway (X-Gateway-Token handshake) — public endpoints are also rate-limited
    per IP at the gateway.'
  title: Go Enterprise API
  version: "1.0"
paths:
...
  /api/v1/gateway/apis:
    get:
      description: Returns all registered APIs managed by the gateway console, with
        live health status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/go-enterprise-api_internal_domain_gatewayapi.GatewayAPI'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List registered gateway APIs
      tags:
      - gateway
    post:
      consumes:
      - application/json
      description: Registers an upstream API and publishes it to the gateway (nginx
        include + hot reload) immediately. base_path must be a single segment like
        /orders; upstream is the origin url. Methods default to GET, rate_limit_rpm
        to 60, requires_auth to true.
      parameters:
      - description: API to register
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler.createGatewayAPIRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_domain_gatewayapi.GatewayAPI'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "409":
          description: base path already in use
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register a gateway API
      tags:
      - gateway
  /api/v1/gateway/apis/{id}:
    delete:
      description: Removes the API from the gateway (generated nginx config is dropped).
      parameters:
      - description: Gateway API ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft delete a registered gateway API
      tags:
      - gateway
    get:
      parameters:
      - description: Gateway API ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_domain_gatewayapi.GatewayAPI'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get one registered gateway API
      tags:
      - gateway
    put:
      consumes:
      - application/json
      description: Updates fields and re-publishes the gateway config. Audited (old
        → new data).
      parameters:
      - description: Gateway API ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler.updateGatewayAPIRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_domain_gatewayapi.GatewayAPI'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a registered gateway API
      tags:
      - gateway
  /api/v1/gateway/apis/{id}/preview:
    get:
      description: Simulates the exact nginx location block the gateway would apply.
      parameters:
      - description: Gateway API ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: generated nginx snippet
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Preview generated nginx config (no publish)
      tags:
      - gateway
  /api/v1/gateway/apis/{id}/restore:
    post:
      description: Re-registers the API and re-publishes the gateway config.
      parameters:
      - description: Gateway API ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore a soft-deleted gateway API
      tags:
      - gateway
  /api/v1/gateway/apis/{id}/stats:
    get:
      description: Today's totals (Redis) plus the most recent proxied requests (in-memory
        ring) for one registered API.
      parameters:
      - description: Gateway API ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interface_http_handler.gatewayStatsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Gateway API statistics
      tags:
      - gateway
  /api/v1/gateway/publish:
    post:
      description: Regenerates every nginx include and triggers the gateway hot-reload.
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go-enterprise-api_internal_interface_httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Re-publish the full registry to the gateway
      tags:
      - gateway
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Provide a JWT access token obtained from /api/v1/auth/login or the
      Entra OIDC flow as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"